api/requirements.txt
flask
mysql-connector
PyJWT